refactor: CategoryFragment.kt argument 제거 -> ViewModel에 저장, money_type -> int로 변경

+++ app/src/main/java/com/example/gagyeboost/common/Const.kt
@@ -8,8 +8,8 @@ const val IS_EXPENSE_KEY = "isExpense"
 const val TODAY_STRING_KEY = "today"
 
 // Money Type
-const val INCOME = 1.toByte()
-const val EXPENSE = 0.toByte()
+const val INCOME = 1
+const val EXPENSE = 0
 
 // HomeFragment, RecordDetailActivity
 const val DATE_DETAIL_ITEM_ID_KEY = "dateDetailItem"
